Rename files for better understanding of files content
